+++ src/modules/application/service/index.ts
@@ -1,0 +1,2 @@
+export * from './application-admin.service';
+export * from './application.service';
