+++ src/modules/user/service/index.ts
@@ -1,0 +1,2 @@
+export * from './user-admin.service';
+export * from './user.service';
