+++ src/modules/user/user.module.ts
@@ -3,8 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { UserController } from './controller/user.controller';
 import { UserRepository } from './repository/user.repository';
-import { UserAdminService } from './service/user-admin.service';
-import { UserService } from './service/user.service';
+import { UserAdminService, UserService } from './service';
 
 @Module({
   imports: [
@@ -19,7 +18,7 @@ import { UserService } from './service/user.service';
     }),
   ],
   controllers: [UserController],
-  providers: [UserService, UserRepository, UserAdminService],
+  providers: [UserRepository, UserService, UserAdminService],
   exports: [],
 })
 export class UserModule {}
